# Use Minio's optimized SHA-256