@buschjost 2020-06-12
I am running with this update merged since over a month and it is running fine for me (on multiple Macbook Pro with Catalina).
It might be a big improvement for some people - I read about TB-size backups that are beeing re-scanned due to interrupted initial backups in the forum.